requirements.txt
Django>=4.2.0,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.8,<2.9
drf-spectacular>=0.27.0,<0.28
geopy>=2.4,<2.5
boto3>=1.34,<1.35
django-storages>=1.14,<1.15